Add offline handling to Checkout button and update dependencies
This commit adds the functionality to prevent a user from checking out when disconnected from the internet. The cart detail component now imports ConnectionService and has a boolean variable 'connected' to keep track of the connection state. The Checkout button is disabled if there are no items in the cart or the connection is off. The text on the button also now reflects this state.

Additionally, the Angular version has been updated in this commit. Several other dependencies have been upgraded too in the package-lock.json file, some have been removed, and some have been added. This update intends to keep the project up to date with the latest suitable dependency versions, ensuring optimal compatibility, performance and security.